[common/meta/types] refactor: move application level error such as UnknownTable, into app_error.rs
