+++ common/meta/api/src/schema_api_impl.rs
@@ -57,36 +57,36 @@ use common_meta_app::schema::UpdateTableMetaReply;
 use common_meta_app::schema::UpdateTableMetaReq;
 use common_meta_app::schema::UpsertTableOptionReply;
 use common_meta_app::schema::UpsertTableOptionReq;
+use common_meta_types::app_error::AppError;
+use common_meta_types::app_error::CreateDatabaseWithDropTime;
+use common_meta_types::app_error::CreateTableWithDropTime;
+use common_meta_types::app_error::DatabaseAlreadyExists;
+use common_meta_types::app_error::DropDbWithDropTime;
+use common_meta_types::app_error::DropTableWithDropTime;
+use common_meta_types::app_error::TableAlreadyExists;
+use common_meta_types::app_error::TableVersionMismatched;
+use common_meta_types::app_error::UndropDbHasNoHistory;
+use common_meta_types::app_error::UndropDbWithNoDropTime;
+use common_meta_types::app_error::UndropTableAlreadyExists;
+use common_meta_types::app_error::UndropTableHasNoHistory;
+use common_meta_types::app_error::UndropTableWithNoDropTime;
+use common_meta_types::app_error::UnknownDatabase;
+use common_meta_types::app_error::UnknownTable;
+use common_meta_types::app_error::UnknownTableId;
 use common_meta_types::txn_condition;
 use common_meta_types::txn_op::Request;
-use common_meta_types::AppError;
 use common_meta_types::ConditionResult;
-use common_meta_types::CreateDatabaseWithDropTime;
-use common_meta_types::CreateTableWithDropTime;
-use common_meta_types::DatabaseAlreadyExists;
-use common_meta_types::DropDbWithDropTime;
-use common_meta_types::DropTableWithDropTime;
 use common_meta_types::MatchSeq;
 use common_meta_types::MatchSeqExt;
 use common_meta_types::MetaError;
 use common_meta_types::MetaId;
 use common_meta_types::Operation;
-use common_meta_types::TableAlreadyExists;
-use common_meta_types::TableVersionMismatched;
 use common_meta_types::TxnCondition;
 use common_meta_types::TxnDeleteRequest;
 use common_meta_types::TxnOp;
 use common_meta_types::TxnOpResponse;
 use common_meta_types::TxnPutRequest;
 use common_meta_types::TxnRequest;
-use common_meta_types::UndropDbHasNoHistory;
-use common_meta_types::UndropDbWithNoDropTime;
-use common_meta_types::UndropTableAlreadyExists;
-use common_meta_types::UndropTableHasNoHistory;
-use common_meta_types::UndropTableWithNoDropTime;
-use common_meta_types::UnknownDatabase;
-use common_meta_types::UnknownTable;
-use common_meta_types::UnknownTableId;
 use common_meta_types::UpsertKVReq;
 use common_proto_conv::FromToProto;
 use common_tracing::tracing;
