+++ common/meta/types/src/lib.rs
@@ -47,6 +47,7 @@ mod user_privilege;
 mod user_quota;
 mod user_stage;
 
+pub mod app_error;
 pub mod error_context;
 mod principal_identity;
 mod proto_display;
@@ -64,6 +65,25 @@ pub mod protobuf {
     pub const FILE_DESCRIPTOR_SET: &[u8] = tonic::include_file_descriptor_set!("meta_descriptor");
 }
 
+pub use app_error::AppError;
+pub use app_error::CreateDatabaseWithDropTime;
+pub use app_error::CreateTableWithDropTime;
+pub use app_error::DatabaseAlreadyExists;
+pub use app_error::DropDbWithDropTime;
+pub use app_error::DropTableWithDropTime;
+pub use app_error::ShareAlreadyExists;
+pub use app_error::TableAlreadyExists;
+pub use app_error::TableVersionMismatched;
+pub use app_error::UndropDbHasNoHistory;
+pub use app_error::UndropDbWithNoDropTime;
+pub use app_error::UndropTableAlreadyExists;
+pub use app_error::UndropTableHasNoHistory;
+pub use app_error::UndropTableWithNoDropTime;
+pub use app_error::UnknownDatabase;
+pub use app_error::UnknownDatabaseId;
+pub use app_error::UnknownShare;
+pub use app_error::UnknownTable;
+pub use app_error::UnknownTableId;
 pub use applied_state::AppliedState;
 pub use change::AddResult;
 pub use change::Change;
@@ -101,27 +121,8 @@ pub use meta_raft_errors::ForwardToLeader;
 pub use meta_raft_errors::MetaRaftError;
 pub use meta_raft_errors::RetryableError;
 pub use meta_result_error::MetaResultError;
-pub use meta_storage_errors::AppError;
-pub use meta_storage_errors::CreateDatabaseWithDropTime;
-pub use meta_storage_errors::CreateTableWithDropTime;
-pub use meta_storage_errors::DatabaseAlreadyExists;
-pub use meta_storage_errors::DropDbWithDropTime;
-pub use meta_storage_errors::DropTableWithDropTime;
 pub use meta_storage_errors::MetaStorageError;
 pub use meta_storage_errors::MetaStorageResult;
-pub use meta_storage_errors::ShareAlreadyExists;
-pub use meta_storage_errors::TableAlreadyExists;
-pub use meta_storage_errors::TableVersionMismatched;
-pub use meta_storage_errors::UndropDbHasNoHistory;
-pub use meta_storage_errors::UndropDbWithNoDropTime;
-pub use meta_storage_errors::UndropTableAlreadyExists;
-pub use meta_storage_errors::UndropTableHasNoHistory;
-pub use meta_storage_errors::UndropTableWithNoDropTime;
-pub use meta_storage_errors::UnknownDatabase;
-pub use meta_storage_errors::UnknownDatabaseId;
-pub use meta_storage_errors::UnknownShare;
-pub use meta_storage_errors::UnknownTable;
-pub use meta_storage_errors::UnknownTableId;
 pub use operation::MetaId;
 pub use operation::Operation;
 pub use principal_identity::PrincipalIdentity;
